Move validation to a dedicated class
